\name{PanicModel-internal}
\title{Internal PanicModel functions}

\alias{d_A_dt}
\alias{d_PT_dt}
\alias{d_H_dt}
\alias{d_E_dt}
\alias{d_AS_dt}
\alias{d_ES_dt}
\alias{d_AV_dt}

\alias{pars_default}
\alias{initial_default}

\description{Internal PanicModel functions.}

\author{Jonas Haslbeck}

\details{

These are internal functions.

}

\keyword{internal}
