\name{PanicModel-internal}
\title{Internal PanicModel functions}

\alias{dA_dt}
\alias{dPT_dt}
\alias{dH_dt}
\alias{dE_dt}
\alias{dS_dt}
\alias{dX_dt}
\alias{dV_dt}

\alias{pars_default}
\alias{initial_default}

\description{Internal PanicModel functions.}

\author{Jonas Haslbeck}

\details{

These are internal functions.

}

\keyword{internal}
